fix(ldt): correct primary key type from int64 to int32
The ID field in LDT files is actually stored as 4-byte i32, not i64.
Update Row struct and related code to use correct type.

+++ src/ldt/csv.rs
@@ -82,7 +82,7 @@ pub fn import_from_csv<R: Read>(reader: &mut R) -> Result<(i32, Vec<FieldDef>, V
         }
 
         // First column is the primary key
-        let primary_key: i64 = record
+        let primary_key: i32 = record
             .get(0)
             .context("Missing primary key column")?
             .parse()
